+++ tests/test_executor.py
@@ -3,7 +3,8 @@
 import pytest
 from mock import patch
 
-from pytest_mysql.executor import MySQLExecutor, MySQLUnsupported
+from pytest_mysql.executor import MySQLExecutor
+from pytest_mysql.exceptions import MySQLUnsupported
 
 
 @pytest.mark.parametrize(